+++ server/schema/deletednsrecordsbulk.sql
@@ -8,7 +8,7 @@ BEGIN
 
 		FOR j IN array_lower(records, 1) .. array_upper(records, 1) LOOP
 			DELETE FROM record USING label INNER JOIN zone ON zone.id = zone_id WHERE label_id = label.id AND zone.name = zones[i]
-			AND label = records[j][2] AND class = records[j][3]::dnsclass AND ttl = records[j][4]::int
+			AND label = records[j][2] AND class = records[j][3]::dnsclass AND ttl = records[j][4]::bigint
 			AND type = records[j][5] AND rdata = records[j][6];
 		END LOOP;
