+++ server/schema/deleteslavezone.sql
@@ -2,7 +2,7 @@ CREATE OR REPLACE FUNCTION DeleteSlaveZone (
 	zonename varchar
 ) RETURNS void AS $$
 DECLARE
-	zone_id_var int;
+	zone_id_var bigint;
 BEGIN
 	SELECT id INTO zone_id_var FROM slavezone WHERE name = zonename;
 	IF NOT FOUND THEN
