+++ server/schema/getdnsrecords.sql
@@ -1,7 +1,7 @@
 CREATE OR REPLACE FUNCTION GetDnsRecords(
 	zonename varchar,
 	labelname varchar,
-	out record_id int,
+	out record_id bigint,
 	out record_label varchar,
 	out record_class dnsclass,
 	out record_ttl int,
